Switch to magick from deprecated magick convert
magick convert has given deprecation warnings for some time now, and
there is a new API, which changes a little. It suffices to reorder the
arguments a little to get equivalent results. As far as I understood,
the arguments get applied in order. So once the input SVG is loaded,
it's immediately rasterized, that's why we need to set the density and
background before the wallpaper.svg. The rest of the options just
applies sequentially to get the same output as before.